code/requirements.txt
h5py==2.10.0
jsonpickle==1.2
jupyter-client==5.3.4
jupyter-core==4.6.1
kornia==0.1.2+1782eb3
matplotlib==3.1.1
ninja==1.10.0.post1
numpy==1.17.3
openmesh==1.1.3
plyfile==0.7.1
progressbar==2.5
pymesh2==0.2.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.2.1
shapely==1.7.0
tensorboard==1.13.1
tensorboardX==1.9
torch==1.3.0
torch-cluster==1.4.5
torch-geometric==1.3.2
torch-scatter==1.3.2
torch-sparse==0.4.3
torchfile==0.1.0
torchvision==0.4.1
tqdm==4.41.1
trimesh==3.2.39
